+++ src/borg/archiver/completion_cmd.py
@@ -714,6 +714,7 @@ class CompletionMixIn:
             "buzhash,19,23,21,4095",
             "buzhash64,19,23,21,4095,2",
             "fastcdc,19,23,21,2",
+            "rabin-aes,19,23,21,2",
         ]
         chunker_params_choices_str = " ".join(chunker_params_choices)